+++ coreengine/storage/note_sqlite_repository.py
@@ -11,6 +11,7 @@ class SqliteNoteRepository(NoteRepository):
     def __serialize_note(self,note:Note)->dict:
         return {
             # note_id is auto incremented
+            'user_id':note.user_id,
             'note_type_id':note.note_type_id,
             'fields_JSON':json.dumps(note.fields,ensure_ascii=False),
             'tags_JSON':json.dumps(note.tags,ensure_ascii=False),
@@ -24,6 +25,7 @@ class SqliteNoteRepository(NoteRepository):
     def __deserialize_note(self,row:sqlite3.Row)->Note:
         return Note(
             note_id=row['note_id'],
+            user_id=row['user_id'],
             note_type_id=row['note_type_id'],
             fields=json.loads(row['fields_JSON']),
             tags=json.loads(row['tags_JSON']),
@@ -34,13 +36,16 @@ class SqliteNoteRepository(NoteRepository):
             hint=row['hint'] if row['hint'] is not None else '',
         )
     
-    def add_note(self,note:Note):
+    def add_note(self, user_id:int, note:Note):
         if note.note_id is not None:
             raise ValueError("Note ID must be None")
+        if not isinstance(user_id,int):
+            raise ValueError("User ID must be an integer")
         data=self.__serialize_note(note)
 
         cursor=self.__conn.execute("""
         INSERT INTO note (
+            user_id,
             note_type_id,
             fields_JSON,
             tags_JSON,
@@ -53,6 +58,7 @@ class SqliteNoteRepository(NoteRepository):
         VALUES (?, ?, ?, ?, ?, ?, ?, ?)
         """,
         (
+            user_id,
             data['note_type_id'],
             data['fields_JSON'],
             data['tags_JSON'],
@@ -65,22 +71,27 @@ class SqliteNoteRepository(NoteRepository):
         )
         return cursor.lastrowid
 
-    def get_note(self,note_id:int)->Note:
+    def get_note(self, user_id:int, note_id:int)->Note:
         if not isinstance(note_id,int):
             raise ValueError("Note ID must be an integer")
+        if not isinstance(user_id,int):
+            raise ValueError("User ID must be an integer")
         row=self.__conn.execute("""
-        SELECT * FROM note WHERE note_id=?
-        """,(note_id,)).fetchone()
+        SELECT * FROM note WHERE note_id=? AND user_id=?
+        """,(note_id,user_id)).fetchone()
         if row is None:
             raise ValueError("Note not found")
         return self.__deserialize_note(row)
     
-    def update_note(self,note:Note):
+    def update_note(self, user_id:int, note:Note):
         if note.note_id is None:
             raise ValueError("Note ID is required")
+        if not isinstance(user_id,int):
+            raise ValueError("User ID must be an integer")
         data=self.__serialize_note(note)
         cursor=self.__conn.execute("""
             UPDATE note SET
+            user_id=?,
             note_type_id=?,
             fields_JSON=?,
             tags_JSON=?,
@@ -90,6 +101,7 @@ class SqliteNoteRepository(NoteRepository):
             hint=?
             WHERE note_id=?
         """,(
+            user_id,
             data['note_type_id'],
             data['fields_JSON'],
             data['tags_JSON'],
@@ -104,13 +116,15 @@ class SqliteNoteRepository(NoteRepository):
             raise ValueError("Note not found")
         return note.note_id
     
-    def delete_note(self,note_id:int):
+    def delete_note(self, user_id:int, note_id:int):
         if not isinstance(note_id, int):
-            raise TypeError("Note id is not an integer")
+            raise ValueError("Note ID must be an integer")
+        if not isinstance(user_id,int):
+            raise ValueError("User ID must be an integer")
 
         cursor = self.__conn.execute(
-            "DELETE FROM note WHERE note_id = ?",
-            (note_id,),
+            "DELETE FROM note WHERE note_id = ? AND user_id = ?",
+            (note_id,user_id),
         )
 
         if cursor.rowcount == 0:
@@ -120,8 +134,17 @@ class SqliteNoteRepository(NoteRepository):
             raise ValueError("Note not found")
         return cursor.rowcount
 
-    def get_all_notes(self):
+    def get_all_notes(self, user_id:int):
+        if not isinstance(user_id,int):
+            raise ValueError("User ID must be an integer")
         rows = self.__conn.execute(
-            "SELECT * FROM note ORDER BY note_id"
+            "SELECT * FROM note WHERE user_id = ? ORDER BY note_id"
+            (user_id,),
         ).fetchall()
         return [self.__deserialize_note(row) for row in rows]
+    
+    def count_notes(self, user_id:int):
+        cursor=self.__conn.execute("""
+        SELECT COUNT(*) FROM note WHERE user_id = ?
+        """,(user_id,))
+        return cursor.fetchone()[0]
